requirements.txt
gcloud==0.18.3
google-cloud-bigquery
pandas
pyarrow